refactor: Add nested enum called Brand in Coke class representing coca-cola and pepsi
